scrubber/requirements.txt
django==3.0
djangorestframework
gunicorn
psycopg2
requests
ipython
ipdb
beautifulsoup4
pandas
lxml
html5lib
scikit-learn
jedi==0.17.2
parso==0.7.1
